+++ react/Row.tsx
@@ -1,16 +1,11 @@
-import React, { FunctionComponent } from 'react'
+import type { PropsWithChildren } from 'react'
+import React from 'react'
 import classNames from 'classnames'
 import { Container } from 'vtex.store-components'
+import type { CssHandlesTypes } from 'vtex.css-handles'
 import { useCssHandles } from 'vtex.css-handles'
 
-import StickyRow from './StickyRow'
-
-interface Props {
-  sticky?: boolean
-  zIndex?: number
-  fullWidth?: boolean
-  inverted?: boolean
-}
+import StickyRow from './components/StickyRow'
 
 const CSS_HANDLES = [
   'headerRowContainer',
@@ -19,14 +14,40 @@ const CSS_HANDLES = [
   'headerRowContentContainer',
 ] as const
 
-const Row: FunctionComponent<Props> = ({
+interface Props {
+  /**
+   * Define if the row behaves in a sticky manner
+   * @default false
+   * */
+  sticky?: boolean
+  zIndex?: number
+  /**
+   * Define if the element takes the full width of its container
+   * @default false
+   * */
+  fullWidth?: boolean
+  /**
+   * Define if the row should have inverted background and text colors
+   * @default false
+   * */
+  inverted?: boolean
+  classes?: CssHandlesTypes.CustomClasses<typeof CSS_HANDLES>
+}
+
+/**
+ * @deprecated Use the `vtex.sticky-layout` app instead.
+ */
+function Row({
   children,
   sticky,
   zIndex,
   fullWidth,
   inverted,
-}) => {
-  const handles = useCssHandles(CSS_HANDLES)
+  classes,
+}: PropsWithChildren<Props>) {
+  const { handles } = useCssHandles(CSS_HANDLES, {
+    classes,
+  })
 
   const content = (
     <div className={`${handles.headerRowContainer} w-100 flex items-center`}>