+++ react/components/StickyRows.tsx
@@ -1,21 +1,29 @@
-import React, { FunctionComponent } from 'react'
+import type { PropsWithChildren } from 'react'
+import React from 'react'
 
 import useCumulativeHeightState from '../hooks/useCumulativeHeightState'
 
-const RowContext = React.createContext<{
+type RowContextType = {
+  /** Callback executed when element is resized */
   onResize(height: number): void
+  /** Vertical Offset from top of container */
   offset: number
-}>({
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
+}
+
+const RowContext = React.createContext<RowContextType>({
   onResize: () => {},
   offset: 0,
 })
 
-const StickyRows: FunctionComponent = ({ children }) => {
+/**
+ * @deprecated Use the `vtex.sticky-layout` app instead
+ */
+// eslint-disable-next-line @typescript-eslint/ban-types
+function StickyRows({ children }: PropsWithChildren<{}>) {
   const { updateRowHeight, getAccumulatedHeight } = useCumulativeHeightState()
 
   return (
-    <React.Fragment>
+    <>
       {React.Children.map(children, (child, index) => (
         <RowContext.Provider
           value={{
@@ -26,9 +34,10 @@ const StickyRows: FunctionComponent = ({ children }) => {
           {child}
         </RowContext.Provider>
       ))}
-    </React.Fragment>
+    </>
   )
 }
 
 export default StickyRows
+
 export { RowContext }