+++ react/legacy/index.d.ts
@@ -1,0 +1,3 @@
+declare const Header: any
+
+export default Header